% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{cuml4r}
\alias{cuml4r}
\title{cuml4r}
\description{
The purpose of 'cuml4r' is to provide a simple and intuitive R interface for
'cuML' (<https://github.com/rapidsai/cuml>). 'cuML' is a suite of GPU-
accelerated machine learning libraries powered by 'CUDA'
(<https://en.wikipedia.org/wiki/CUDA>).
}
\author{
Yitao Li <yitao@rstudio.com>
}
